backend/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
Markdown==3.3.3
psycopg2-binary==2.8.6
requests==2.25.0
